server/requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
blake3==1.0.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.15.0
colorama==0.4.6
dataclasses-json==0.6.7
datasets==2.14.4
deprecated==1.2.18
dill==0.3.7
dirtyjson==1.0.8
distro==1.9.0
et-xmlfile==2.0.0
evaluate==0.4.3
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
greenlet==3.2.1
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
llama-cloud==0.1.15
llama-cloud-services==0.6.6
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-openai==0.3.1
llama-index-extractors-entity==0.3.1
llama-index-graph-stores-neo4j==0.4.6
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-openai==0.3.26
llama-index-llms-openai-like==0.3.4
llama-index-llms-openllm==0.4.1
llama-index-multi-modal-llms-openai==0.4.3
llama-index-postprocessor-cohere-rerank==0.3.0
llama-index-postprocessor-flag-embedding-reranker==0.3.0
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-readers-smart-pdf-loader==0.3.0
llama-parse==0.6.4.post1
llmsherpa==0.1.4
macholib==1.16.3
markupsafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.15
mypy-extensions==1.1.0
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.76.0
openpyxl==3.1.5
orjson==3.10.16
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pip==24.3.1
platformdirs==4.3.7
polars==1.28.1
propcache==0.3.1
psutil==7.0.0
pyarrow==19.0.1
pydantic==2.11.3
pydantic-core==2.33.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
pymupdf==1.25.4
pypdf==5.4.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
seqeval==1.2.2
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
span-marker==1.7.0
sqlalchemy==2.0.40
starlette==0.46.2
striprtf==0.0.26
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
trio==0.30.0
trio-websocket==0.12.2
types-requests==2.32.0.20250328
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
websocket-client==1.8.0
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
./wheels/doc_search-0.1.0-py3-none-any.whl